BigDecimal bevezetese a DrawnBlock szamara a kurzor csak a Precision-nak megfelelo poziciokat veheti fel Meg dolgozni kell rajta

+++ trunk/src/hu/akoel/mgu/drawnblock/DrawnBlockFactory.java
@@ -1,9 +1,11 @@
 package hu.akoel.mgu.drawnblock;
 
+import java.math.BigDecimal;
+
 import hu.akoel.mgu.drawnblock.DrawnBlock.Status;
 
 public interface DrawnBlockFactory {
 
-	public DrawnBlock getNewDrawnBlock( Status status, double x1, double y1 );
+	public DrawnBlock getNewDrawnBlock( Status status, BigDecimal x1, BigDecimal y1 );
 	
 }
